client: refactor names, adjust tests
